add AuthRestController
