+++ src/main/resources/templates/fragments/authorization-button.html
@@ -52,7 +52,7 @@
                     </div>
                 </div>
                 <div class="action-root_align-center-mUlAP">
-                    <a href="#"
+                    <a id="buttonAuth"
                        class="btn btn-md auth-form-auth-form__submit-rSWaC"
                        style="background: #0af"
                        role="button">Войти</a>
